Changed the type of field Id in Item domain to BigInteger

+++ src/main/java/com/mcompany/inventory/model/Counter.java
@@ -3,24 +3,26 @@ package com.mcompany.inventory.model;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
+import java.math.BigInteger;
+
 @Document(collection = "counters")
 public class Counter {
 
     @Id
     private String id;
 
-    private int seq;
+    private BigInteger seq;
 
-    public Counter(String id, int seq) {
+    public Counter(String id, BigInteger seq) {
         this.id = id;
         this.seq = seq;
     }
 
-    public void setSeq(int inc) {
+    public void setSeq(BigInteger inc) {
         this.seq = inc;
     }
 
-    public int getSeq() {
+    public BigInteger getSeq() {
         return this.seq;
     }
