+++ src/main/java/com/mcompany/inventory/service/DbRepository.java
@@ -1,5 +1,6 @@
 package com.mcompany.inventory.service;
 
+import java.math.BigInteger;
 import java.util.List;
 
 
@@ -8,6 +9,6 @@ interface DbRepository<T>  {
     void create(T object);
     void update(T object);
     void delete(int id);
-    T findOne(int id);
+    T findOne(BigInteger id);
     List<T> findAll();
 }
